tools/requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.4.0
requests>=2.32.3

